requirements.txt
cffi==1.15.0
click==8.1.3
cryptography==37.0.2
Flask==2.1.2
Flask-Cors==3.0.10
greenlet==1.1.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
marshmallow==3.16.0
packaging==21.3
pycparser==2.21
PyJWT==2.4.0
PyMySQL==1.0.2
pyparsing==3.0.9
python-dotenv==0.20.0
six==1.16.0
SQLAlchemy==1.4.37
Werkzeug==2.1.2
